refactor(cloudformation): move the RDS types into RdsCfnProvisioner
Moves all seven AWS::RDS types out of the monolith switch into a per-service
provisioner. Method bodies are carried over unchanged except for the six
master-credential call sites described below; the rest is plumbing.

The seven types delete by physical id alone, so the id-only delete override
serves all of them and none appears in DELETE_NEEDS_STACK_RESOURCE.

Unlike the ECS, ELBv2, Cognito, EC2-networking, Events and Batch batches, this
one does not take its service out of the monolith. SecretTargetAttachment is
still served there and reads an instance's or cluster's endpoint through
RdsService to build its connection detail, so the field, its constructor
parameter and its import all stay. Four helpers the RDS block shared with other
types stay too, copied into the provisioner: sameNameExistingResource,
deleteRenamedResource, parseIntProp and parseCfnTags. firstNonBlank,
parseBoolProp and parseOptionalIntProp had no non-RDS callers left and moved
outright.

Name generation now goes through ProvisionContext.generatePhysicalName rather
than a copy: the two implementations were already byte-identical, down to the
12-character suffix.

MasterUsername and MasterUserPassword keep the opt-out #3301 gave them. The
engine resolves dynamic references for every property now, with ssm-secure
rejected, so those two properties read through
resolveOptionalWithoutDynamicReferences and resolve their own references with
ssm-secure allowed. Reading them with the ordinary resolveOptional compiles and
leaves the suite green while breaking ssm-secure on the only resource AWS
permits it on, so the monolith's helper moves across with the code rather than
collapsing into the ordinary path.

CloudFormationRdsDynamicReferenceIntegrationTest is new and covers that, because
nothing did. RdsCfnProvisionerTest builds its engine without a dynamic-reference
resolver, so the general stage is off there and its assertions hold either way,
and CloudFormationDynamicReferenceIntegrationTest has no RDS case. The ssm-secure
test fails with ROLLBACK_COMPLETE when the opt-out is reverted, and the plain ssm
test guards the other direction.

The 69 existing RdsCfnProvisionerTest tests and the 2 integration tests pass
unchanged, which is the guard that the move preserved behaviour: both build
through CfnProvisionerFixture, so they exercise the new provisioner without
knowing it exists. The fixture's CfnDynamicReferences wiring had to move ahead
of the registry build, since the provisioner takes it as a constructor argument.

Making the types provisioner-owned rather than LEGACY_SWITCH puts them under
CfnSchemaCoverageTest for the first time, which the switch was exempt from. The
19 attributes it found are recorded as gaps here, unchanged from what the switch
set. Two of them are closable and a follow-up PR closes them, kept out of this one
so it stays a pure move; Endpoint and
ReadEndpoint are recorded rather than set because the dotted sub-attributes
already carry them, which is the form AWS's GetAtt actually exposes.

Monolith 5,942 to 5,458 lines, legacy types 39 to 32.
