+++ src/test/resources/cloudformation/supported-resource-types.tsv
@@ -109,13 +109,13 @@ AWS::Organizations::OrganizationalUnit	OrganizationsCfnProvisioner
 AWS::Organizations::Policy	OrganizationsCfnProvisioner
 AWS::Organizations::ResourcePolicy	OrganizationsCfnProvisioner
 AWS::Pipes::Pipe	PipesCfnProvisioner
-AWS::RDS::DBCluster	LEGACY_SWITCH
-AWS::RDS::DBClusterParameterGroup	LEGACY_SWITCH
-AWS::RDS::DBInstance	LEGACY_SWITCH
-AWS::RDS::DBParameterGroup	LEGACY_SWITCH
-AWS::RDS::DBProxy	LEGACY_SWITCH
-AWS::RDS::DBProxyTargetGroup	LEGACY_SWITCH
-AWS::RDS::DBSubnetGroup	LEGACY_SWITCH
+AWS::RDS::DBCluster	RdsCfnProvisioner
+AWS::RDS::DBClusterParameterGroup	RdsCfnProvisioner
+AWS::RDS::DBInstance	RdsCfnProvisioner
+AWS::RDS::DBParameterGroup	RdsCfnProvisioner
+AWS::RDS::DBProxy	RdsCfnProvisioner
+AWS::RDS::DBProxyTargetGroup	RdsCfnProvisioner
+AWS::RDS::DBSubnetGroup	RdsCfnProvisioner
 AWS::Route53::HostedZone	Route53CfnProvisioner
 AWS::Route53::RecordSet	LEGACY_SWITCH
 AWS::S3::Bucket	S3CfnProvisioner
